Rename module typing to types
